KP3285 and KP-3462 complete

+++ kairosActivity/src/main/java/com/kairos/activity/persistence/repository/activity/CustomActivityMongoRepository.java
@@ -46,5 +46,5 @@ public interface CustomActivityMongoRepository {
 
     List<ActivityDTO> findAllActivitiesWithTimeTypesByUnit(Long unitId);
 
-    Set<Long> findAllActivitiesByUnitIdAndUnavailableTimeType(long unitId);
+    Set<BigInteger> findAllActivitiesByUnitIdAndUnavailableTimeType(long unitId);
 }
